+++ web/profiles/custom/os2loop/modules/os2loop_documents/modules/os2loop_documents_fixtures/src/Fixture/DocumentFixture.php
@@ -197,6 +197,56 @@ BODY,
       $this->setReference($document->getType() . ':' . $document->getTitle(), $document);
       $document->save();
     }
+
+    $document = Node::create([
+      'type' => 'os2loop_documents_document',
+      'title' => 'Dokument om cykler',
+      'os2loop_documents_document_autho' => 'Egon',
+      'os2loop_shared_subject' => [
+        'target_id' => $this->getReference('os2loop_subject:Diverse')->id(),
+      ],
+      'os2loop_shared_tags' => [
+        ['target_id' => $this->getReference('os2loop_tag:test')->id()],
+      ],
+    ]);
+
+    $paragraph = Paragraph::create([
+      'type' => 'os2loop_documents_text_and_image',
+      'os2loop_documents_tai_text' => [
+        'value' => <<<'BODY'
+Når man kører på cykel møder man ofte andre cykler så derfor skal der være styr på cyklens grej.
+BODY,
+        'format' => 'os2loop_documents_rich_text',
+      ],
+    ]);
+    $paragraph->save();
+    $document->get('os2loop_documents_document_conte')->appendItem($paragraph);
+    $document->save();
+
+    $document = Node::create([
+      'type' => 'os2loop_documents_document',
+      'title' => 'Dokument om cykler',
+      'os2loop_documents_document_autho' => 'Egon',
+      'os2loop_shared_subject' => [
+        'target_id' => $this->getReference('os2loop_subject:Diverse')->id(),
+      ],
+      'os2loop_shared_tags' => [
+        ['target_id' => $this->getReference('os2loop_tag:test')->id()],
+      ],
+    ]);
+
+    $paragraph = Paragraph::create([
+      'type' => 'os2loop_documents_text_and_image',
+      'os2loop_documents_tai_text' => [
+        'value' => <<<'BODY'
+Der skal være styr på cyklens grej.
+BODY,
+        'format' => 'os2loop_documents_rich_text',
+      ],
+    ]);
+    $paragraph->save();
+    $document->get('os2loop_documents_document_conte')->appendItem($paragraph);
+    $document->save();
   }
 
   /**